Port rspace engine — RSpace play/replay + SpaceMatcher matching
Completes the rspace critical path (Law 8/11):
- RSpace (two-step-locked consume/produce/install + checkpoint/soft-checkpoint)
- ReplayRSpace (replay trace bookkeeping: rig/checkReplayData)
- SpaceMatcher matching + i64 datum-index type correction

+++ crates/rspace/src/internal.rs
@@ -75,7 +75,7 @@ pub struct ConsumeCandidate<C, A> {
     pub channel: C,
     pub datum: Datum<A>,
     pub removed_datum: A,
-    pub datum_index: usize,
+    pub datum_index: i64,
 }
 
 /// A matched continuation candidate during produce (port of `ProduceCandidate`).
